% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datFET.donor2}
\alias{datFET.donor2}
\title{Micro RNA Expression Values for the Fetal Human Brain Atlas.}
\format{A data frame with 9919840 rows and 4 variables:
\describe{
 \item{donorID}{brain ID from the Allen Institute}
 \item{gene}{gene the above probe corresponds to; gene to probe is many to one}
 \item{brain_structure}{brain structure acronym}
 \item{value}{normalized expression value}
}}
\source{
\url{"http://www.brainspan.org/static/download.html"}
}
\usage{
datFET.donor2
}
\description{
A dataset containing the normalized expression values for a single brain in the fetal human brain atlas.
}
\keyword{datasets}
