% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donorsFET}
\alias{donorsFET}
\title{Donor ID's for Brains in the Fetal Human Brain Atlas}
\format{A character vector of length 4:}
\source{
\url{"http://www.brainspan.org/static/download.html"}
}
\usage{
donorsFET
}
\description{
A vector of the donor ID's from the 4 brains in the fetal human brain atlas
}
\keyword{datasets}
