% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genesFET}
\alias{genesFET}
\title{Genes for Brains in the Human Brain Atlas}
\format{A character vector of length 29176:}
\source{
\url{"http://www.brainspan.org/static/download.html"}
}
\usage{
genesFET
}
\description{
A vector of the genes measured from the 4 brains in the fetal human brain atlas
}
\keyword{datasets}
