% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{structuresFET}
\alias{structuresFET}
\title{Brain structures measured in the Fetal Human Brain}
\format{A character vector of length 516:}
\source{
\url{"http://www.brainspan.org/static/download.html"}
}
\usage{
structuresFET
}
\description{
A vector of the brain structure acronyms measured in the fetal human brain atlas
}
\keyword{datasets}
